Cleanup Message model
