+++ app/models/message.rb
@@ -1,52 +1,28 @@
 class Message < ApplicationRecord
+  include DocumentImage, Cancellable
+
   belongs_to :assistant
   belongs_to :conversation
   belongs_to :content_document, class_name: "Document", optional: true
   belongs_to :run, optional: true
 
-  has_many :documents, dependent: :destroy
-
   enum role: %w[user assistant].index_by(&:to_sym)
 
   delegate :user, to: :conversation
 
-  accepts_nested_attributes_for :documents
-
   before_validation :set_default_role, on: :create
-  before_validation :create_conversation, on: :create, if: -> { conversation.blank? }
+  before_validation :create_conversation, on: :create, if: -> { conversation.blank? }, prepend: true
 
   validates :role, presence: true
   validates :content_text, presence: true, unless: :assistant?
-  validate :validate_conversation, if: -> { conversation.present? && Current.user }
-  validate :validate_assistant, if: -> { assistant.present? && Current.user }
+  validate  :validate_conversation,  if: -> { conversation.present? && Current.user }
+  validate  :validate_assistant,     if: -> { assistant.present? && Current.user }
 
   scope :ordered, -> { order(:created_at) }
 
   after_create :start_assistant_reply, if: -> { user? }
 
   after_save :update_assistant_on_conversation, if: -> { assistant.present? && conversation.present? }
-  after_save :save_cancelled_id_to_redis, if: :saved_change_to_cancelled_at?
-
-  def has_document_image?(variant = nil)
-    has_image = documents.present? && documents.first.file.attached?
-    if has_image && variant
-      has_image = documents.first.has_file_variant_processed?(variant)
-    end
-
-    !!has_image
-  end
-
-  def document_image_path(variant, fallback: nil)
-    return nil unless has_document_image?
-
-    if documents.first.has_file_variant_processed?(variant)
-      documents.first.fully_processed_url(variant)
-    elsif fallback.nil?
-      documents.first.redirect_to_processed_path(variant)
-    else
-      fallback
-    end
-  end
 
   private
 
@@ -74,15 +50,5 @@ class Message < ApplicationRecord
   def update_assistant_on_conversation
     return if conversation.assistant == assistant
     conversation.update!(assistant: assistant)
-  end
-
-  def save_cancelled_id_to_redis
-    redis.set("message-cancelled-id", id)
-  end
-
-  private
-
-  def redis
-    RedisConnection.client
   end
 end
